Rename ghosts.rs to construction.rs
