+++ emergence_lib/src/structures/construction.rs
